refactor: Split out the list integration tests.
